Update main.py
An almost working No fish eye effect raycaster. However, completing the raycaster without the rewriting it seems way too inefficient, and i think im going to rewrite the entire raycasting code with the fisheye effect in mind. 
It would be easier to test out something which I searched up about the fish eye effect, where the end of each ray has equal spacing, which would be difficult, but not such a mess like the current code.